Towards eddy
Change from eddy to eddy_openmp which was recently downloaded.
Eddy_openmp sent a dimension error, the mask wasn't the same size as the scan. The c3d tool was erased and I tried to pass the number of slices that were cropped from the b0_P and rb0_A in Extract_b0s_FSL to ProcessingDTI_FSL in order to crop the XinFile as well, but the spm_write_vol doesn't work for 4dimensional scans